Rearranged and doxy'ed code for the checklistController.
At this point, should be easily buildable.  Interface should be complete, though internal work is going to change significantly, the header file should be near a final
 version.

Now open to separate implementation of the MFD using the MFDConnector and the information provided in the checklistController.h file to provide proper instructions.