added comments  and removed debug strings
all modes appear to be working. was able to set manual position in lunar orbit with the controls, and then reacquire automaticially at earthrise

I have tested REACQ mode through 5 lunar orbits and it works flawlessly on my end.

Scan-limit warning and scan limit are working now. the antenna switches back to wide mode after hitting the warning and switches back to manual(in reaqc mode). in auto mode the antenna "stops tracking" onb scanlimit which is correct per CSM Systems handbook.

The switch, tracking mode, and scanlimit logic, and the actual tracking control logic are now seperated. tracking control could probably be a function, but it would need 12 arguments so I've left it for now.