Don't rely on configurator to close sockets
Before this commit the PCM used some DLL hackery to clsoe the telemetry
sockets on sim exit. Why this was done is beyond me but I assume it was
for legacy functionality that is now removed. This also fixes the
occasional "bind failed" error because the LM socket was stored as the
CM socket.